+++ src/client/connection_like.rs
@@ -1,17 +1,19 @@
 use std::borrow::Cow;
 
 use async_trait::async_trait;
-use futures::future::BoxFuture;
+use futures::{future::BoxFuture, FutureExt};
 use rmpv::Value;
 use serde::de::DeserializeOwned;
 
-use super::{Executor, Stream, Transaction, TransactionBuilder};
+use super::Executor;
 use crate::{
     codec::{
-        request::{Call, Delete, Eval, Insert, Ping, Replace, RequestBody, Select, Update, Upsert},
+        request::{
+            Call, Delete, EncodedRequest, Eval, Insert, Ping, Replace, Request, Select, Update,
+            Upsert,
+        },
         utils::deserialize_non_sql_response,
     },
-    errors::Error,
     IteratorType, Result,
 };
 
@@ -21,30 +23,13 @@ pub trait ConnectionLike: Executor {
     ///
     /// It is not recommended to use this method directly, since some requests
     /// should be only sent in specific situations and might break connection.
-    #[deprecated = "This method will be removed in future"]
-    fn send_any_request<R>(&self, body: R) -> BoxFuture<Result<Value>>
+    fn send_request<R>(&self, body: R) -> BoxFuture<Result<Value>>
     where
-        R: RequestBody;
-
-    /// Get new [`Stream`].
-    ///
-    /// It is safe to create `Stream` from any type, implementing current trait.
-    fn stream(&self) -> Stream;
-
-    /// Prepare [`TransactionBuilder`], which can be used to override parameters and create
-    /// [`Transaction`].
-    ///
-    /// It is safe to create `TransactionBuilder` from any type.
-    fn transaction_builder(&self) -> TransactionBuilder;
-
-    /// Create [`Transaction`] with parameters from builder.
-    ///
-    /// It is safe to create `Transaction` from any type, implementing current trait.
-    async fn transaction(&self) -> Result<Transaction>;
+        R: Request;
 
     /// Send PING request ([docs](https://www.tarantool.io/en/doc/latest/dev_guide/internals/box_protocol/#iproto-ping-0x40)).
     async fn ping(&self) -> Result<()> {
-        self.send_any_request(Ping {}).await.map(drop)
+        self.send_request(Ping {}).await.map(drop)
     }
 
     // TODO: docs
@@ -53,7 +38,7 @@ pub trait ConnectionLike: Executor {
         I: Into<Cow<'static, str>> + Send,
         T: DeserializeOwned,
     {
-        let body = self.send_any_request(Eval::new(expr, args)).await?;
+        let body = self.send_request(Eval::new(expr, args)).await?;
         deserialize_non_sql_response(body).map_err(Into::into)
     }
 
@@ -63,9 +48,7 @@ pub trait ConnectionLike: Executor {
         I: Into<Cow<'static, str>> + Send,
         T: DeserializeOwned,
     {
-        let body = self
-            .send_any_request(Call::new(function_name, args))
-            .await?;
+        let body = self.send_request(Call::new(function_name, args)).await?;
         deserialize_non_sql_response(body).map_err(Into::into)
     }
 
@@ -83,7 +66,7 @@ pub trait ConnectionLike: Executor {
         T: DeserializeOwned,
     {
         let body = self
-            .send_any_request(Select::new(
+            .send_request(Select::new(
                 space_id, index_id, limit, offset, iterator, keys,
             ))
             .await?;
@@ -93,7 +76,7 @@ pub trait ConnectionLike: Executor {
     // TODO: docs
     // TODO: decode response
     async fn insert(&self, space_id: u32, tuple: Vec<Value>) -> Result<()> {
-        let _ = self.send_any_request(Insert::new(space_id, tuple)).await?;
+        let _ = self.send_request(Insert::new(space_id, tuple)).await?;
         Ok(())
     }
 
@@ -108,7 +91,7 @@ pub trait ConnectionLike: Executor {
         tuple: Vec<Value>,
     ) -> Result<()> {
         let _ = self
-            .send_any_request(Update::new(space_id, index_id, index_base, keys, tuple))
+            .send_request(Update::new(space_id, index_id, index_base, keys, tuple))
             .await?;
         Ok(())
     }
@@ -124,7 +107,7 @@ pub trait ConnectionLike: Executor {
         tuple: Vec<Value>,
     ) -> Result<()> {
         let _ = self
-            .send_any_request(Upsert::new(space_id, index_base, ops, tuple))
+            .send_request(Upsert::new(space_id, index_base, ops, tuple))
             .await?;
         Ok(())
     }
@@ -132,7 +115,7 @@ pub trait ConnectionLike: Executor {
     // TODO: structured tuple
     // TODO: decode response
     async fn replace(&self, space_id: u32, keys: Vec<Value>) -> Result<()> {
-        let _ = self.send_any_request(Replace::new(space_id, keys)).await?;
+        let _ = self.send_request(Replace::new(space_id, keys)).await?;
         Ok(())
     }
 
@@ -140,30 +123,19 @@ pub trait ConnectionLike: Executor {
     // TODO: decode response
     async fn delete(&self, space_id: u32, index_id: u32, keys: Vec<Value>) -> Result<()> {
         let _ = self
-            .send_any_request(Delete::new(space_id, index_id, keys))
+            .send_request(Delete::new(space_id, index_id, keys))
             .await?;
         Ok(())
     }
 }
 
 #[async_trait]
-impl<C: ConnectionLike + super::private::Sealed + Sync> ConnectionLike for &C {
-    fn send_any_request<R>(&self, body: R) -> BoxFuture<Result<Value>>
+impl<E: Executor + ?Sized> ConnectionLike for E {
+    fn send_request<R>(&self, body: R) -> BoxFuture<Result<Value>>
     where
-        R: RequestBody,
+        R: Request,
     {
-        (*self).send_any_request(body)
-    }
-
-    fn stream(&self) -> Stream {
-        (*self).stream()
-    }
-
-    fn transaction_builder(&self) -> TransactionBuilder {
-        (*self).transaction_builder()
-    }
-
-    async fn transaction(&self) -> Result<Transaction> {
-        (*self).transaction().await
+        let req = EncodedRequest::new(body, None);
+        async move { (*self).send_encoded_request(req?).await }.boxed()
     }
 }